requirements.txt
langchain
langchain-google-genai
langchain-community
langchain-text-splitters
langchain-classic
langchain-huggingface
chromadb
pypdf
streamlit
python-dotenv
sentence-transformers
